% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_vpc_to_params}
\alias{append_vpc_to_params}
\title{Append VPC Values to Params Data Frame}
\usage{
append_vpc_to_params(params, vpcfunc, x)
}
\arguments{
\item{params}{A data frame where each row contains parameter values, including
the necessary variables like \code{b0}, \code{b1}, \code{sig11}, \code{sig12}, \code{sig22}, and \code{phi}.}

\item{vpcfunc}{A function that calculates the VPC, taking \code{beta}, \code{Sigma}, \code{phi},
and \code{x} as arguments.}

\item{x}{A numeric vector. Each element in \code{x} will be used to generate
a VPC column in the \code{params} data frame.}
}
\value{
A modified version of the \code{params} data frame with new columns for
each VPC calculated based on the \code{x} values.
}
\description{
This function calculates Variance Partition Coefficient (VPC) values for
each row in the \code{params} data frame and appends the results as new columns.
Each value of \code{x} generates a corresponding VPC column.
}
