% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{compareGLMMFit}
\alias{compareGLMMFit}
\title{Compare GLMM Fit for True and False Families}
\usage{
compareGLMMFit(
  params,
  covariates,
  X,
  ns,
  true_family,
  false_family,
  formula,
  iter = 1,
  seed = NULL
)
}
\arguments{
\item{params}{A matrix or data frame of parameter values for GLMM data
generation.}

\item{covariates}{A vector or matrix of covariates for the GLMM.}

\item{X}{A design matrix for the GLMM.}

\item{ns}{Number of samples or observations per batch.}

\item{true_family}{The family distribution assumed to be correct.}

\item{false_family}{An alternative family distribution to compare.}

\item{formula}{The model formula for GLMM fitting.}

\item{iter}{Number of iterations for data generation (default is 1).}

\item{seed}{An optional seed for reproducibility.}
}
\value{
A list containing the coefficients of the true family, false family,
and VST-transformed model (if applicable).
}
\description{
Compare GLMM Fit for True and False Families
}
