% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{comparePoints}
\alias{comparePoints}
\title{Compare Estimated VPC with True VPC}
\usage{
comparePoints(
  params,
  ns,
  X = X,
  family,
  link,
  formula,
  vpc_input_values = NULL,
  iter = 1,
  seed = NULL,
  num_cores = 1
)
}
\arguments{
\item{params}{A parameter matrix used for generating the data.
Columns may include parameters like \code{b0}, \code{b1}, \code{sig11},
\code{sig12}, \code{sig22}.}

\item{ns}{A vector specifying the sample sizes for each group in the data
generation.}

\item{X}{The covariate matrix for fixed effects. Default is \code{X = X}.}

\item{family}{A character string specifying the family for the model
(e.g., "negative_binomial", "tweedie").}

\item{link}{The link function to be used in the GLMM model.}

\item{formula}{The formula for fitting the GLMM model (e.g.,
\code{Feature ~ X + (X|cluster)}).}

\item{vpc_input_values}{A vector of input values at which to compute the VPC.
Default is \code{NULL}.}

\item{iter}{Number of iterations for generating data. Default is 1.}

\item{seed}{An optional random seed for reproducibility. Default is
\code{NULL}.}

\item{num_cores}{The number of CPU cores to use for parallel processing.
Default is 1.}
}
\value{
A list with two elements: \code{true} and \code{vpcest}.
\itemize{
\item \code{true}: The true VPC values based on the input parameters.
\item \code{vpcest}: The estimated VPC values from the fitted GLMM models.
The function also stores the number of iterations as an attribute called
\code{num_iter}.
}
}
\description{
This function compares the true Variance Partition Coefficient (VPC) values
calculated from parameter matrices with the estimated VPC values from GLMM
model fits.
}
