% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{customboxplot}
\alias{customboxplot}
\title{Custom Boxplot for Variance Partition Coefficients (VPCs)}
\usage{
customboxplot(x, ...)
}
\arguments{
\item{x}{A vpcestpt object containing VPC estimates and true VPC values.}

\item{...}{Additional arguments to be passed to the custom_boxplot function.}
}
\description{
This function generates boxplots of variance partition coefficients (VPCs) from a given
vpcestpt object. It overlays the true VPC values on the boxplots for comparison.
}
