% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{generatecompareEstimation}
\alias{generatecompareEstimation}
\title{Compare GLMM Fit for True and False Families}
\usage{
generatecompareEstimation(
  params,
  ns,
  gen_family,
  another_family,
  gen_link,
  fit_formula,
  X = NULL,
  vpc_input_values = NULL,
  num_cores = 1,
  fit_transform = TRUE,
  iter = 1,
  seed = NULL
)
}
\arguments{
\item{params}{A matrix or data frame of parameter values for GLMM data
generation.}

\item{ns}{Number of samples or observations per batch.}

\item{gen_family}{The family distribution used for data generation.}

\item{another_family}{An alternative family distribution to compare.}

\item{gen_link}{A link function for generating the data}

\item{fit_formula}{The model formula for GLMM fitting.}

\item{X}{A design matrix for the GLMM.}

\item{vpc_input_values}{A vector for which vpc is calculated}

\item{num_cores}{Number of cores to fit GLMM with}

\item{fit_transform}{logical wether a VST model should be fitter and its vpc
computed.}

\item{iter}{Number of iterations for data generation (default is 1).}

\item{seed}{An optional seed for reproducibility.}
}
\value{
A list containing the coefficients of the true family, false family,
and VST-transformed model (if applicable).
}
\description{
Compare GLMM Fit for True and False Families
}
