% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{generatecompareGroups}
\alias{generatecompareGroups}
\title{Compare GLMM Fit for Groups when handled independently}
\usage{
generatecompareGroups(
  params,
  ns,
  family,
  link,
  formula,
  X = NULL,
  vpc_input_values = NULL,
  num_cores = 1,
  iter = 1,
  seed = NULL
)
}
\arguments{
\item{params}{A matrix or data frame of parameter values for GLMM data
generation.}

\item{ns}{Number of samples or observations per batch.}

\item{family}{The family distribution used for data generation.}

\item{link}{A link function for generating the data}

\item{formula}{The model formula for GLMM fitting.}

\item{X}{A design matrix for the GLMM.}

\item{vpc_input_values}{A vector for which vpc is calculated}

\item{num_cores}{Number of cores to fit GLMM with}

\item{iter}{Number of iterations for data generation (default is 1).}

\item{seed}{An optional seed for reproducibility.}
}
\value{
A list containing the coefficients of the true vpcs, vpcs with
single mixed effect model, and vpcs for independent mixed models.
}
\description{
Compare GLMM Fit for Groups when handled independently
}
