% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paramgridWithVPC}
\alias{paramgridWithVPC}
\title{Generate Parameter Grid with VPC Values}
\usage{
paramgridWithVPC(b0s, b1s, sigmas, phis)
}
\arguments{
\item{b0s}{A vector of values for the intercept term \code{b0}.}

\item{b1s}{A vector of values for the slope term \code{b1}.}

\item{sigmas}{A list of sigma matrices (with each element containing sig11,
sig12, sig22).}

\item{phis}{A vector of \code{phi} values for the model.}
}
\value{
A data frame containing the parameter grid with VPC values appended.
}
\description{
This function generates a parameter grid based on inputs of \code{b0s}, \code{b1s},
\code{sigmas}, and \code{phis}, and then appends Variance Partition Coefficient (VPC)
values for each row based on specified values of \code{x}.
}
\examples{
b0s <- c(3, 7)
b1s <- c(-5, 5)
sigmas <- list(c(2, 1, 2), c(2, -1, 2), c(2, 0, 4))
phis <- seq(0.01, 1, 0.05)
params_with_vpc <- paramgridWithVPC(b0s, b1s, sigmas, phis)
}
