% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percomPlot}
\alias{percomPlot}
\title{Generate bias plots for multiple models using various vpc columns}
\usage{
percomPlot(true_params, ..., palette_name = "Set1")
}
\arguments{
\item{true_params}{A data frame containing the true values for vpc columns. The columns should be named using the prefix 'vpc', for example, \code{vpc0}, \code{vpc1}, etc.}

\item{...}{One or more data frames containing the model estimates for vpc columns to be compared against \code{true_params}.}

\item{palette_name}{A string specifying the name of the \code{RColorBrewer} palette to be used. The default is \code{"Set1"}. Use \code{RColorBrewer::display.brewer.all()} to see available palettes.}
}
\value{
Generates plots for each model's vpc columns. The function does not return any value.
}
\description{
This function generates bias plots for each model's vpc columns compared to the true values.
It allows customization of the color palette using \code{RColorBrewer}.
}
