% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.vpcestgr}
\alias{plot.vpcestgr}
\title{This function creates a 2x2 grid of plots for visualizing various aspects of vpcestgr objects.}
\usage{
\method{plot}{vpcestgr}(x, ...)
}
\arguments{
\item{x}{An object of class vpcestgr}

\item{...}{Additional arguments passed to plot}
}
\value{
No return value, called for side effects (plotting)
}
\description{
This function creates a 2x2 grid of plots for visualizing various aspects of vpcestgr objects.
}
