% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_counts}
\alias{transform_counts}
\title{Transform Count Data}
\usage{
transform_counts(ys, fun, num_cores)
}
\arguments{
\item{ys}{A matrix or data frame containing the count data to be transformed.}

\item{fun}{A function to be applied to the count data. This function should
accept the count data as its first argument
and optionally additional parameters like \code{num_cores} for parallel execution.}

\item{num_cores}{number of PC cores}
}
\value{
The transformed count data as returned by the function \code{fun}.
If the transformation fails, the function
returns \code{NULL} and an error message is printed.
}
\description{
This function takes a matrix of count data and applies a transformation
function (e.g., variance stabilizing transformation)
to it. The transformation is executed with error handling in case the
transformation fails.
}
\examples{
ys <- matrix(rpois(20, lambda = 10), ncol = 5)
transform_counts(ys, vst_transform)
}
