% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paramgridWithVPC}
\alias{paramgridWithVPC}
\title{Generate Parameter Grid with VPC Values}
\usage{
paramgridWithVPC(b0s, b1s, sigmas, phis, vpc_func, power = NULL, ...)
}
\arguments{
\item{b0s}{A vector of values for the intercept term \code{b0}.}

\item{b1s}{A vector of values for the slope term \code{b1}.}

\item{sigmas}{A list of sigma matrices (with each element containing sig11,
sig12, sig22).}

\item{phis}{A vector of \code{phi} values for the model.}

\item{vpc_func}{The VPC function that computes VPC values.}

\item{power}{A vector of \code{power} values for the model.}

\item{...}{Additional arguments for the VPC function.}
}
\value{
A data frame containing the parameter grid with VPC values appended.
}
\description{
This function generates a parameter grid based on inputs of \code{b0s}, \code{b1s},
\code{sigmas}, and \code{phis}, and then appends Variance Partition Coefficient (VPC)
values for each row based on specified values of \code{x}.
}
